cove-ai-core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pgvector==0.2.5
orjson==3.10.7
